pyproject.toml
[project]
name = "debate-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.14",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
